chore: rename AssertTemplate's misspelling
